bigintegerfield

+++ defacto/models.py
@@ -6,8 +6,8 @@ class SupplyDemand(models.Model):
     name = models.CharField(max_length=50)
     code = models.CharField(max_length=6)
     # volume = models.IntegerField(blank=True, null=True)
-    institution_possession = models.IntegerField(blank=True, null=True)
-    foreigner_possession = models.IntegerField(blank=True, null=True)
+    institution_possession = models.BigIntegerField(blank=True, null=True)
+    foreigner_possession = models.BigIntegerField(blank=True, null=True)
     institution_average_price = models.FloatField(blank=True, null=True)
     foreigner_average_price = models.FloatField(blank=True, null=True)
